% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.kerdiest-package}
\alias{mlr3learners.kerdiest}
\alias{mlr3learners.kerdiest-package}
\title{mlr3learners.kerdiest: Learners from the {kerdiest} package for 'mlr3'}
\description{
Adds learners from the {kerdiest} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
